Splitting out delete from cron file
